abstract implementation to accept any MediaQueryList object
