+++ packages/web-components/fast-components/src/card/card.styles.ts
@@ -2,15 +2,7 @@ import { css } from "@microsoft/fast-element";
 import { display, elevation } from "../styles";
 import { SystemColors } from "../styles/system-colors";
 import { neutralLayerCardBehavior } from "../styles/recipes";
-import { HighContrastStylesheetBehavior } from "../styles/high-contrast-stylesheet-behavior";
-
-export const HighConstrastCardStyles = css`
-    :host {
-        forced-color-adjust: none;
-        border: calc(var(--outline-width) * 1px) solid ${SystemColors.CanvasText};
-        background: ${SystemColors.Canvas};
-    }
-`;
+import { forcedColorsStylesheetBehavior } from "../styles/match-media-stylesheet-behavior";
 
 export const CardStyles = css`
     ${display("block")} :host {
@@ -26,5 +18,13 @@ export const CardStyles = css`
     }
 `.withBehaviors(
     neutralLayerCardBehavior,
-    new HighContrastStylesheetBehavior(HighConstrastCardStyles)
+    forcedColorsStylesheetBehavior(
+        css`
+            :host {
+                forced-color-adjust: none;
+                border: calc(var(--outline-width) * 1px) solid ${SystemColors.CanvasText};
+                background: ${SystemColors.Canvas};
+            }
+        `
+    )
 );
